Initial Checkin
Demo page for a friend who wanted to do some specific text to HTML
table processing, so I used this as an excuse to try the HTML5 drag and
drop and file APIs and provide a solution that generates a HTML table
on the fly using JavsScript.

The file input.txt is sample input, the file basic.txt is the original
Acorn BASIC file I was provided by a friend.